+++ .changeset/cool-planets-decide.md
@@ -1,0 +1,5 @@
+---
+"@frontify/guideline-blocks-settings": patch
+---
+
+refactor (Toolbar): split Toolbar into smaller subcomponents
