+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/AttachmentsToolbarButton/AttachmentsToolbarButtonTrigger.tsx
@@ -3,7 +3,7 @@
 import { IconCaretDown12, IconPaperclip16 } from '@frontify/fondue';
 import { type AttachmentsTriggerProps } from '../../../Attachments/types';
 
-import { BaseToolbarButton } from './BaseToolbarButton';
+import { BaseToolbarButton } from '../BaseToolbarButton';
 
 export const AttachmentsToolbarButtonTrigger = ({ children, isFlyoutOpen }: AttachmentsTriggerProps) => (
     <BaseToolbarButton forceActiveStyle={isFlyoutOpen}>
