+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/BaseToolbarButton.tsx
@@ -1,7 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { type HTMLAttributes, type ReactNode, forwardRef } from 'react';
-import { getToolbarButtonClassNames } from '../helpers';
+import { getToolbarButtonClassNames } from './helpers';
 
 type BaseToolbarButtonProps = {
     children: ReactNode;
