+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/DragHandleToolbarButton/DragHandleToolbarButton.tsx
@@ -2,7 +2,7 @@
 
 import { DEFAULT_DRAGGING_TOOLTIP, DEFAULT_DRAG_TOOLTIP } from '../..';
 import { useIsDragPreview } from '../context/DragPreviewContext';
-import { BaseToolbarButton } from './BaseToolbarButton';
+import { BaseToolbarButton } from '../BaseToolbarButton';
 import { ToolbarButtonTooltip } from '../ToolbarButtonTooltip';
 
 export type DragHandleToolbarButtonProps = {
