+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/MenuToolbarButton/index.ts
@@ -1,3 +1,3 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-export * from './ToolbarButton';
+export * from './MenuToolbarButton';
