+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/index.ts
@@ -1,4 +1,8 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './Toolbar';
+export * from './AttachmentsToolbarButton';
+export * from './DragHandleToolbarButton';
+export * from './FlyoutToolbarButton';
+export * from './MenuToolbarButton';
 export * from './types';
